+++ dev/apps/converters/ash2mdp.cpp
@@ -111,7 +111,7 @@ public:
       if (debugLevel>4)
          MDPHeader::hexDump = true;
 
-      for (int i=0; i<clipOption.getCount(); i++)
+      for (size_t i=0; i<clipOption.getCount(); i++)
       {
          string s=clipOption.getValue()[i];
          int sep = s.find_first_not_of("0123456789kKMG");
