+++ dev/apps/visibility/svvis.cpp
@@ -207,7 +207,7 @@ bool SVVis::initialize(int argc, char *argv[]) throw()
       // get the ephemeris source(s)
    ephReader.verboseLevel = verboseLevel;
    FFIdentifier::debugLevel = debugLevel;
-   for (int i=0; i<ephFileOpt.getCount(); i++)
+   for (size_t i=0; i<ephFileOpt.getCount(); i++)
       ephReader.read(ephFileOpt.getValue()[i]);
    if (ephReader.eph == NULL)
    {
@@ -425,7 +425,7 @@ void SVVis::process()
    if (riseSet)
    {
       RiseSetList::const_iterator rsli;
-      for (int prn = 1; prn < prs.size(); prn++)
+      for (size_t prn = 1; prn < prs.size(); prn++)
       {
          cout << setw(2) << left << prn;
          const RiseSetList& rsl = prs[prn];
